Pin dependency versions
Unpinned dependency versions can cause build issues in Nix if the
package registry index snapshot contains a newer version than what is
packaged in nixpkgs.

See nixos/nixpkgs#43430 for context.

The versions chosen here are the most recent ones currently packaged
in nixpkgs, with the exception of ibrowse which has been pinned to a
version that is not affected by cmullaparthi/ibrowse#160.